[WIP] Clean up abandoned WAL directories
Remove WAL directories that are no longer associated with a `ManagedInstance`
that this agent is responsible for. The storage directories used by active
instances are compared to all subdirectories under the `wal_directory` root.
Any not associated with an instance and that haven't been modified in over a
configured amount of time are removed.